requirements.txt
accelerate==0.21.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
attrs==24.2.0
bitsandbytes==0.44.0
blessed==1.20.0
bypy==1.8.5
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
datasets==3.0.1
decorator==5.1.1
deepspeed==0.12.6
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
einops==0.6.1
einops-exts==0.0.4
exceptiongroup==1.2.2
fastapi==0.115.0
ffmpy==0.4.0
filelock==3.16.1
flash-attn 
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
gpustat==1.1.1
gradio==5.0.2
gradio_client==1.4.0
h11==0.14.0
h5py==3.12.1
hjson==3.1.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
importlib_resources==6.4.5
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
latex2mathml==3.77.0
-e git+https://github.com/haotian-liu/LLaVA.git@c121f0432da27facab705978f83c4ada465e46fd#egg=llava
llvmlite==0.43.0
markdown-it-py==3.0.0
markdown2==2.5.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
moviepy==2.1.1
mpmath==1.3.0
ms-swift==2.5.0.dev0
multidict==6.1.0
multiprocess==0.70.16
narwhals==1.8.4
networkx==3.3
ninja==1.11.1.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
openai==0.28.0
opencv-python==4.10.0.84
orjson==3.10.7
packaging==24.1
pandas==2.2.3
peft==0.13.0
pillow==10.4.0
platformdirs==4.3.6
proglog==0.1.10
protobuf==5.28.2
psutil==6.0.0
py-cpuinfo==9.0.0
pyarrow==17.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pynvml==11.5.3
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.8.1
rpds-py==0.20.0
ruff==0.6.8
safetensors==0.4.5
scikit-learn==1.2.2
scipy==1.14.1
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==2.14.0
setproctitle==1.3.3
shellingham==1.5.4
shortuuid==1.0.13
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
speedtest-cli==2.1.3
starlette==0.38.6
svgwrite==1.4.3
sympy==1.13.3
threadpoolctl==3.5.0
timm==0.6.13
tokenizers==0.15.2
tomlkit==0.12.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.5
transformers==4.37.2
triton==2.1.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.0
wandb==0.18.2
wavedrom==2.0.3.post3
wcwidth==0.2.13
websockets==11.0.3
xxhash==3.5.0
yarl==1.13.1
peft